+++ lib/strutil/strutilascii.c
@@ -540,11 +540,11 @@ finally:
 
 /* --------------------------------------------------------------------------------------------- */
 
-static int
+static size_t
 str_ascii_offset_to_pos (const char *text, size_t length)
 {
     (void) text;
-    return (int) length;
+    return length;
 }
 
 /* --------------------------------------------------------------------------------------------- */
